// src/serverApi/v3/models/card-element-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { CardRichTextElementResponse } from './card-rich-text-element-response';

/**
 * 
 * @export
 * @interface CardElementResponse
 */
export interface CardElementResponse {
    /**
     * The id of the card element
     * @type {string}
     * @memberof CardElementResponse
     */
    id: string;
    /**
     * Type of card element
     * @type {string}
     * @memberof CardElementResponse
     */
    cardElementType: CardElementResponseCardElementType;
    /**
     * Content of the card element, depending on its type
     * @type {CardRichTextElementResponse}
     * @memberof CardElementResponse
     */
    content: CardRichTextElementResponse;
}

/**
    * @export
    * @enum {string}
    */
export enum CardElementResponseCardElementType {
    RICH_TEXT = 'richText'
}