// src/serverApi/v3/models/card-element-params.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { RichTextCardElementParam } from './rich-text-card-element-param';

/**
 * 
 * @export
 * @interface CardElementParams
 */
export interface CardElementParams {
    /**
     * 
     * @type {string}
     * @memberof CardElementParams
     */
    id?: string;
    /**
     * Content of the card element, depending on its type
     * @type {RichTextCardElementParam}
     * @memberof CardElementParams
     */
    content: RichTextCardElementParam;
}